requirements.txt
alembic @ file:///tmp/build/80754af9/alembic_1637680647426/work
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
autopage @ file:///home/conda/feedstock_root/build_artifacts/autopage_1642834347039/work
certifi==2021.10.8
charset-normalizer==2.0.10
clearml==1.1.6
cliff @ file:///home/conda/feedstock_root/build_artifacts/cliff_1638109549784/work
cmaes @ file:///home/conda/feedstock_root/build_artifacts/cmaes_1613785714721/work
cmd2 @ file:///D:/bld/cmd2_1640528825633/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
colorlog==5.0.1
furl==2.1.3
future==0.18.2
greenlet @ file:///C:/ci/greenlet_1628888257991/work
hanging-threads==2.0.5
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.4.0
joblib==1.1.0
jsonschema==4.4.0
Mako @ file:///tmp/build/80754af9/mako_1610650590832/work
MarkupSafe @ file:///C:/ci/markupsafe_1621528383308/work
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186163140/work
mkl-service==2.4.0
numpy==1.21.5
olefile==0.46
optuna @ file:///home/conda/feedstock_root/build_artifacts/optuna_1633337702246/work
orderedmultidict==1.0.1
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.3.5
pathlib2==2.3.6
pbr @ file:///tmp/build/80754af9/pbr_1619600803852/work
Pillow==9.0.0
prettytable @ file:///home/conda/feedstock_root/build_artifacts/prettytable_1641549356281/work
psutil==5.9.0
PyJWT==2.1.0
pyparsing==3.0.7
pyperclip @ file:///home/conda/feedstock_root/build_artifacts/pyperclip_1622337600177/work
pyreadline==2.1
pyrsistent==0.18.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
PyYAML==6.0
requests==2.27.1
scikit-learn==1.0.2
scipy @ file:///C:/ci/scipy_1641555141383/work
six @ file:///tmp/build/80754af9/six_1623709665295/work
SQLAlchemy @ file:///C:/ci/sqlalchemy_1638290815491/work
stevedore @ file:///D:/bld/stevedore_1639542811759/work
threadpoolctl==3.0.0
torch==1.10.1
torchaudio==0.10.1
torchsummary==1.5.1
torchvision==0.11.2
tqdm @ file:///tmp/build/80754af9/tqdm_1635330843403/work
typing_extensions==4.0.1
urllib3==1.26.8
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
wincertstore==0.2
yappi @ file:///D:/bld/yappi_1636301943748/work
zipp @ file:///opt/conda/conda-bld/zipp_1641824620731/work
